+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 CHANGELOG
 =========
 
-2.3, 2023-10-28
+2.3, 2023-10-29
 ---------------
 - add command-line interface, with export, import, parse, search and stats
 - add option to rename SQL window pages
 - add support for CREATE TABLE .. STRICT (Python 3.9+)
+- include read-only pragmas in PRAGMA SQL
 - update SQLite keywords from 126 to the latest 147-element list
 - use a custom lockfile implementation for single instance check in Linux
 - fix updating main page upon removing databases when home-row is also selected